Use details/fwd.hpp with forward declarations. Cleanup of type conversions, casts and other elements detected during strict compilation (Visual C++ options /W4 /Za).
--HG--
extra : convert_revision : svn%3A1766ff46-f334-0410-ab20-d63176f87757/trunk%40280

+++ src/detail/writer10.cpp
@@ -92,7 +92,7 @@ void WriterImpl::WriteHeader(LASHeader const& header)
     detail::write_n(m_ofs, n4, sizeof(n4));
 
     // 16. Point Data Format ID
-    n1 = header.GetDataFormatId();
+    n1 = static_cast<uint8_t>(header.GetDataFormatId());
     detail::write_n(m_ofs, n1, sizeof(n1));
 
     // 17. Point Data Record Length
